On the 'move-tracking-2' branch: Add casts where converting revnums to 'int', to make a nice clean build.
Found by: brane

* subversion/libsvn_delta/branch.c
  (svn_branch_repos_find_el_rev_by_path_rev): Add a cast.

* subversion/svnmover/svnmover.c
  (svn_branch_find_predecessor_el_rev): Add a cast.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/move-tracking-2@1664441 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svnmover/svnmover.c
@@ -1332,7 +1332,7 @@ svn_branch_find_predecessor_el_rev(svn_branch_el_rev_id_t *old_el_rev,
     return NULL;
 
   branch_id = svn_branch_instance_get_id(old_el_rev->branch, result_pool);
-  rev_root = svn_array_get(repos->rev_roots, old_el_rev->rev - 1);
+  rev_root = svn_array_get(repos->rev_roots, (int)(old_el_rev->rev - 1));
   branch = svn_branch_revision_root_find_branch_by_id(rev_root, branch_id,
                                                       result_pool);
